some fixes and stuff

+++ Source/Engine.Graphics/DX11/Bindings/IndexBuffer.h
@@ -4,12 +4,12 @@
 class IndexBuffer : public Bindable
 {
 public:
-    IndexBuffer(Graphics& gfx, const std::vector<unsigned short>& indices);
-    IndexBuffer(Graphics& gfx, std::string tag, const std::vector<unsigned short>& indices);
+    IndexBuffer(Graphics& gfx, const std::vector<unsigned int>& indices);
+    IndexBuffer(Graphics& gfx, std::string tag, const std::vector<unsigned int>& indices);
     void Bind(Graphics& gfx) noexcept override;
     UINT GetCount() const noexcept;
     static std::shared_ptr<IndexBuffer> Resolve(Graphics& gfx, const std::string& tag,
-        const std::vector<unsigned short>& indices);
+        const std::vector<unsigned int>& indices);
     template<typename...Ignore>
     static std::string GenerateUID(const std::string& tag, Ignore&&...ignore)
     {
